requirements.txt
aiogram==3.*
aiosqlite==0.20.*
httpx==0.27.*
selectolax==0.3.*
python-dotenv==1.0.*
apscheduler==3.10.*
openai==1.*
pydantic==2.*
